Split Rust in several files
