+++ rust/src/buffer.rs
@@ -1,0 +1,26 @@
+use std::slice;
+use std::str;
+
+
+#[no_mangle]
+#[repr(C)]
+#[derive(Copy, Clone, Debug)]
+pub struct Buffer {
+    pub size: usize,
+    pub bytes: *const u8,
+}
+
+impl Buffer {
+    pub fn null() -> Self {
+        Self{size: 0, bytes: std::ptr::null()}
+    }
+    pub fn from_bytes(data: &[u8]) -> Self {
+        Self{size: data.len(), bytes: data.as_ptr()}
+    }
+    pub fn from_str(data: &str) -> Self {
+        Self::from_bytes(data.as_bytes())
+    }
+    pub unsafe fn to_str(&self) -> &str {
+        str::from_utf8_unchecked(slice::from_raw_parts(self.bytes as *const _, self.size))
+    }
+}
